[refactor] itemContent 분리 및 bottom Btn 수정
