+++ client/src/Components/ListItem/itemContent.js
@@ -17,7 +17,6 @@ function ItemContent({ info, screen, navigation }) {
   const description = info.letter ? info.letter : info.description;
 
   // status : wait, process, done, cancel, default
-  const status = info.status ? info.status : 'default';
   const leftBtnText = screen === 'branding' ? '더보기' : '취소됨';
   const rightBtnText = screen === 'branding' ? '대표 등록' : '시술 완료';
 
@@ -28,42 +27,6 @@ function ItemContent({ info, screen, navigation }) {
         text: '삭제하기',
         onPress: () => {
           deleteSubmitHandler(id);
-        },
-      },
-    ]);
-  };
-  const cancelAlert = (id) => {
-    Alert.alert('정말 취소하시겠습니까?', '취소후에는 변경이 불가능합니다', [
-      { text: '취소', style: 'cancel' },
-      {
-        text: '취소하기',
-        onPress: () => {
-          statusSubmitHandler(id, 'cancel');
-        },
-      },
-    ]);
-  };
-  const doneAlert = (bid) => {
-    Alert.alert('시술이 완료되었습니까?', '완료후에는 변경이 불가능합니다', [
-      { text: '취소', style: 'cancel' },
-      {
-        text: '완료하기',
-        onPress: () => {
-          matchingHistoryHandler(bid);
-        },
-      },
-    ]);
-  };
-  const moveToDetailBranding = (info) => {
-    navigation.navigate('DetailBranding', { info });
-  };
-  const registerAlert = (id) => {
-    Alert.alert('대표 포트폴리오로 등록하시겠습니까?', '', [
-      { text: '취소', style: 'cancel' },
-      {
-        text: '등록하기',
-        onPress: () => {
-          registerSubmitHandler(id);
         },
       },
     ]);
@@ -91,140 +54,45 @@ function ItemContent({ info, screen, navigation }) {
         console.error(error);
       });
   };
-  const registerSubmitHandler = async (id) => {
-    await fetch('http://127.0.0.1:3000' + `/api/branding/main`, {
-      method: 'PATCH',
-      headers: {
-        'Content-Type': 'application/json;charset=UTF-8',
-      },
-      body: JSON.stringify({
-        id: id,
-        user_id: info.user_id,
-      }),
-    })
-      .then((response) => response.json())
-      .then(async (response) => {
-        if (response) {
-          Alert.alert('대표 포트폴리오 설정되었습니다!');
-          navigation.push('BrandingMain');
-        }
-      })
-      .catch((error) => {
-        console.error(error);
-      });
-  };
-  const statusSubmitHandler = async (id, status) => {
-    await fetch('http://127.0.0.1:3000' + `/api/bid/status/${id}`, {
-      method: 'PATCH',
-      headers: {
-        'Content-Type': 'application/json;charset=UTF-8',
-      },
-      body: JSON.stringify({
-        status,
-      }),
-    })
-      .then((response) => response.json())
-      .then(async (response) => {
-        if (response) {
-          Alert.alert('Bid 상태가 성공적으로 수정되었습니다!');
-          navigation.push('BidMain', { screen: 'ProcessBidList' });
-        }
-      })
-      .catch((error) => {
-        console.error(error);
-      });
-  };
-  const matchingHistoryHandler = async (bid) => {
-    const { id, customer_id, designer_id, proposal_id } = bid;
-    await fetch('http://127.0.0.1:3000' + `/api/matchingHistory/register`, {
-      method: 'POST',
-      headers: {
-        'Content-Type': 'application/json;charset=UTF-8',
-      },
-      body: JSON.stringify({
-        bid_id: id,
-        customer_id: bid.user.id,
-        designer_id: bid.designer_id,
-        proposal_id: bid.proposal.id,
-      }),
-    })
-      .then((response) => response.json())
-      .then(async (response) => {
-        if (response) {
-          Alert.alert('Bid 상태가 성공적으로 수정되었습니다!');
-          navigation.dispatch(CommonActions.navigate('ProcessBidList'));
-        }
-      })
-      .catch((error) => {
-        console.error(error);
-      });
-  };
 
   return (
     <View style={styles.container}>
-      <View style={styles.bidBox}>
-        <ItemHeader
-          navigation={navigation}
-          info={info}
-          screen={screen}
-          setModalVisible={setModalVisible}
-        />
-        <View style={styles.bidContentArea}>
-          <View style={styles.bidProfileArea}>
-            <Image source={{ uri: img_src }} style={styles.bidProfileImg} />
+      <View style={styles.profileBox}>
+        <Image source={{ uri: img_src }} style={styles.profileImg} />
+      </View>
+      <View style={styles.bidInfoArea}>
+        <View style={styles.nameArea}>
+          <Text style={styles.nameText}>{title}</Text>
+        </View>
+        <View style={styles.locationArea}>
+          <View style={styles.locationView}>
+            <Ionicons name="at" size={15} />
+            <Text style={styles.locationText}>{address}</Text>
           </View>
-          <View style={styles.bidInfoArea}>
-            <View style={styles.nameArea}>
-              <Text style={styles.nameText}>{title}</Text>
-            </View>
-            <View style={styles.locationArea}>
-              <View style={styles.locationView}>
-                <Ionicons name="at" size={15} />
-                <Text style={styles.locationText}>{address}</Text>
-              </View>
-              <View style={styles.locationView}>
-                <Ionicons
-                  name={screen === 'branding' ? 'md-cut-outline' : 'location-outline'}
-                  size={15}
-                />
-                <Text style={styles.locationText}>{distance_limit}</Text>
-              </View>
-            </View>
-            <View style={styles.tagArea}>
-              {keywords &&
-                keywords.length > 0 &&
-                keywords.map((keyword, index) => (
-                  <View style={styles.tagView} key={index}>
-                    <Text style={styles.tagText}># {keyword}</Text>
-                  </View>
-                ))}
-            </View>
-            <View style={styles.descriptionArea}>
-              <Text style={styles.descriptionText} numberOfLines={2}>
-                {description}
-              </Text>
-            </View>
+          <View style={styles.locationView}>
+            <Ionicons
+              name={screen === 'branding' ? 'md-cut-outline' : 'location-outline'}
+              size={15}
+            />
+            <Text style={styles.locationText}>{distance_limit}</Text>
           </View>
         </View>
+        <View style={styles.tagArea}>
+          {keywords &&
+            keywords.length > 0 &&
+            keywords.map((keyword, index) => (
+              <View style={styles.tagView} key={index}>
+                <Text style={styles.tagText}># {keyword}</Text>
+              </View>
+            ))}
+        </View>
+        <View style={styles.descriptionArea}>
+          <Text style={styles.descriptionText} numberOfLines={2}>
+            {description}
+          </Text>
+        </View>
       </View>
-      {status !== 'wait' && (
-        <BottomButton
-          info={info}
-          navigation={navigation}
-          btnDisable={status === 'cancel' || status == 'done' ? true : false}
-          leftBtnText={leftBtnText}
-          leftBtnHandler={() => {
-            screen === 'branding' ? moveToDetailBranding(info) : cancelAlert(info.id);
-          }}
-          rightBtnText={rightBtnText}
-          rightBtnHandler={() => {
-            screen === 'branding' ? registerAlert(info.id) : doneAlert(info);
-          }}
-          status={status}
-        />
-      )}
 
-      <View style={styles.line}></View>
       <Modal
         animationType="slide"
         transparent={true}
@@ -251,27 +119,23 @@ function ItemContent({ info, screen, navigation }) {
 const styles = StyleSheet.create({
   container: {
     flex: 1,
-  },
-  bidBox: {
-    flex: 1,
     padding: 16,
     paddingBottom: 0,
-  },
-
-  bidContentArea: {
-    flexDirection: 'row',
-    marginTop: 16,
     marginBottom: 16,
     alignItems: 'center',
+    flexDirection: 'row',
   },
 
-  bidProfileArea: {
+  profileBox: {
     marginRight: 16,
   },
-  bidProfileImg: {
-    width: 110,
-    height: 110,
+  profileImg: {
+    width: 120,
+    height: 120,
     resizeMode: 'cover',
+  },
+  infoBox: {
+    flex: 1,
   },
   nameArea: {
     marginBottom: 8,
@@ -317,7 +181,7 @@ const styles = StyleSheet.create({
     fontWeight: '400',
   },
   descriptionArea: {
-    width: '80%',
+    width: '60%',
   },
   descriptionText: {
     color: '#111111',
@@ -325,10 +189,6 @@ const styles = StyleSheet.create({
     lineHeight: 16,
   },
 
-  line: {
-    height: 10,
-    backgroundColor: '#f4f4f4',
-  },
   modalContainer: {
     width: '100%',
     height: '15%',
@@ -360,9 +220,6 @@ const styles = StyleSheet.create({
     color: '#111111',
     fontSize: 16,
     lineHeight: 20,
-  },
-  bidInfoArea: {
-    flex: 1,
   },
 });
