+++ client/src/Screen/02_SecondTab/Designer/waitBidListScreen.js
@@ -1,6 +1,8 @@
 import React, { useState, useEffect } from 'react';
 import { StyleSheet, Text, View, ScrollView, Image } from 'react-native';
+import ItemHeader from '../../../Components/ListItem/itemHeader';
 import ItemContent from '../../../Components/ListItem/itemContent';
+import Line from '../../../Components/Common/line';
 
 function WaitBidListScreen({ navigation, bidList }) {
   const [waitBidList, setWaitBidList] = useState([]);
@@ -11,7 +13,11 @@ function WaitBidListScreen({ navigation, bidList }) {
   return (
     <ScrollView style={styles.container}>
       {waitBidList.map((bid, index) => (
-        <ItemContent key={index} info={bid} navigation={navigation} screen="bid" />
+        <View style={styles.bidContainer} key={index}>
+          <ItemHeader navigation={navigation} info={bid} screen="bid" />
+          <ItemContent info={bid} navigation={navigation} screen="bid" />
+          <Line />
+        </View>
       ))}
     </ScrollView>
   );
@@ -21,6 +27,7 @@ const styles = StyleSheet.create({
   container: {
     flex: 1,
   },
+  bidContainer: {},
 });
 
 export default WaitBidListScreen;
